extract components from cart page
