+++ src/pages/CartPage/CartItem/CartItem.css
@@ -1,0 +1,38 @@
+.cart-item-container {
+  height: 64px;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-bottom: 16px;
+}
+
+.cart-item-image {
+  height: 64px;
+  width: 64px;
+}
+
+.cart-item-description {
+  display: flex;
+  flex-direction: column;
+  margin-left: 16px;
+  width: 260px;
+}
+
+.cart-item-description-title {
+  font-size: 14px;
+  color: #171e27;
+  width: 100%;
+}
+
+.cart-item-description-subtitle {
+  font-size: 14px;
+  color: #9b9b9b;
+  width: 100%;
+}
+
+.cart-item-price {
+  font-size: 18px;
+  flex-grow: 1;
+  text-align: right;
+}
